+++ lib/screens/notes_screen.dart
@@ -4,6 +4,7 @@ import 'package:provider/provider.dart';
 
 import '../l10n/app_localizations.dart';
 import '../l10n/labels.dart';
+import '../models/money.dart';
 import '../models/note.dart';
 import '../models/transaction_filter.dart' show foldForSearch;
 import '../providers/settings_provider.dart';
@@ -256,7 +257,7 @@ class _NoteTile extends StatelessWidget {
             ? l10n.noteFilterOverdue
             : DateFormat.yMMMEd(l10n.localeName).format(due),
       if (category != null) category.label(l10n),
-      if (note.amount != null) currency.format(note.amount!.toDouble()),
+      if (note.amount != null) currency.money(note.amount!),
     ];
 
     return Dismissible(